Refactor MPC library type conversion and configuration
