+++ clients/desktop/src/vault/utils/storageVault.ts
@@ -1,13 +1,12 @@
 import { create } from '@bufbuild/protobuf';
 import { Timestamp, TimestampSchema } from '@bufbuild/protobuf/wkt';
+import { fromLibType, toLibType } from '@core/communication/utils/libType';
 import {
   Vault,
   Vault_KeyShareSchema,
   VaultSchema,
 } from '@core/communication/vultisig/vault/v1/vault_pb';
-import { defaultMpcLib } from '@core/mpc/config';
-import { mpcLibName } from '@core/mpc/mpcLibName';
-import { mirrorRecord } from '@lib/utils/record/mirrorRecord';
+import { defaultMpcLib } from '@core/mpc/mpcLib';
 import { convertDuration } from '@lib/utils/time/convertDuration';
 
 import { storage } from '../../../wailsjs/go/models';
@@ -25,7 +24,7 @@ export const toStorageVault = ({
   keyShares,
   localPartyId,
   resharePrefix,
-  libType = defaultMpcLib,
+  libType = toLibType(defaultMpcLib),
 }: Vault): storage.Vault => ({
   name: name,
   public_key_ecdsa: publicKeyEcdsa,
@@ -44,7 +43,7 @@ export const toStorageVault = ({
   order: 0,
   is_backed_up: true,
   coins: [],
-  lib_type: mpcLibName[libType],
+  lib_type: fromLibType(libType),
   convertValues: () => {},
 });
 
@@ -87,5 +86,5 @@ export const fromStorageVault = (vault: storage.Vault): Vault =>
       })
     ),
     resharePrefix: vault.reshare_prefix,
-    libType: mirrorRecord(mpcLibName)[vault.lib_type],
+    libType: toLibType(vault.lib_type ?? defaultMpcLib),
   });
