--- core/mpc/config.ts
@@ -1,3 +1,0 @@
-import { LibType } from "@core/communication/vultisig/keygen/v1/lib_type_message_pb";
-
-export const defaultMpcLib: LibType = LibType.GG20;
