+++ core/mpc/mpcLib.ts
@@ -1,0 +1,5 @@
+export const mpcLibs = ["GG20", "DKLS"];
+
+export type MpcLib = (typeof mpcLibs)[number];
+
+export const defaultMpcLib: MpcLib = "GG20";
